// src/shared/enums/transaction-return-type.enum.ts
export enum TransactionReturnTypeEnum {
   CUS_PROVIDED_LABEL_BACK_TO_THEM = 1,
   CUS_PROVIDED_LABEL_TO_CLIENT = 2,
   CUS_WILL_PICKUP = 3,
   SHIP_BACK_TO_CUS_AND_CHARGE_FEE = 4,
   SHIP_BACK_TO_CUS_AND_REQ_LABEL = 5,
   SHIP_TO_CLIENT_OR_ENTER_ADDRESS_AND_CHARGE_FEE = 6,
   COUNTY_MAILED_REG_OR_PLATE_TO_DRIVER = 7,
   DRIVER_PICKED_UP_REG_OR_PLATE_AT_COUNTY_AND_NO_FEE = 8
}