// src/shared/enums/seller-info.enum.ts
export enum SellerTypeEnum {
    GADEALER = "ga_dealer",
    OUT_OF_STATE_DEALER = "out_of_state_dealer",
    OUT_OF_STATE_TRANSFER = "out_of_state_transfer",
    BIZ_IND_NON_DEL = "business_individual_non_dealer",
    BIZ_NON_DEALER = "business_non_dealer",
    IND_NON_DEALER = "individual_non_dealer",
    GOV = "government"
}